refactor(members): functional member module + grantable members.manage
Replace the MemberService class with plain exported functions (matching the
sibling member-queries.ts style), split across invitations, accept-invitation,
member-mutations, guards, and invitation-links modules.

Swap the hardcoded isAdminOrOwner service-layer gate for the grantable
members.manage capability (requireMemberManage), keeping role-relative rank
guards (OWNER > ADMIN > USER) so a grantee manages USER-tier members without
acting on ADMIN/OWNER or minting authority above their own rank. Closes the
members-enforcement follow-up from the settings-admin-areas work.

Convert TRPCError throws to AuxxError subclasses so lib code drops its
@trpc/server dependency, and remove the router try/catch wrappers that would
otherwise mask an AuxxError as a 500. Update all call sites (member/organization
routers, admin-service, email/permissions, entity-processors, seed) and the two
upload test mocks.
