+++ packages/lib/src/members/index.ts
@@ -1,2 +1,19 @@
-export { findMemberByUser, isAdminOrOwner, listMembersWithUser } from './member-queries'
-export { MemberService } from './member-service'
+export { acceptInvitation, acceptInvitationById } from './accept-invitation'
+export {
+  cancelInvitation,
+  getInvitationLink,
+  getMyPendingInvitations,
+  getPendingInvitations,
+  inviteMember,
+  resendInvitation,
+} from './invitations'
+export { removeMember, updateMemberRole, updateMemberSeatType } from './member-mutations'
+export {
+  findMemberByUser,
+  getActiveMemberCount,
+  getMembership,
+  getOrganizationMembers,
+  isAdminOrOwner,
+  isMember,
+  listMembersWithUser,
+} from './member-queries'
