+++ packages/lib/src/members/member-queries.ts
@@ -1,8 +1,8 @@
 // packages/lib/src/members/member-queries.ts
 
 import { type Database, database as defaultDb, schema } from '@auxx/database'
-import type { OrganizationMemberInfo } from '@auxx/database/types'
-import { and, eq } from 'drizzle-orm'
+import type { OrganizationMemberInfo, OrganizationRole } from '@auxx/database/types'
+import { and, asc, count, eq } from 'drizzle-orm'
 import { getOrgCache } from '../cache'
 
 /** Find membership for a user within an organization (uses org cache) */
@@ -90,6 +90,99 @@ export async function listMembersWithUser(
       (r.user?.name || '').toLowerCase().includes(needle) ||
       (r.user?.email || '').toLowerCase().includes(needle)
   ) as any
+}
+
+/** Get the membership record for a user in an organization. */
+export async function getMembership(
+  userId: string,
+  organizationId: string,
+  db: Database = defaultDb
+) {
+  return db.query.OrganizationMember.findFirst({
+    where: (om, { and, eq }) => and(eq(om.userId, userId), eq(om.organizationId, organizationId)),
+  })
+}
+
+/** Check whether a user is a (real, non-system) member of the organization. */
+export async function isMember(
+  userId: string,
+  organizationId: string,
+  db?: Database
+): Promise<boolean> {
+  // If a specific db instance is passed (e.g. inside a transaction), query directly
+  if (db && db !== defaultDb) {
+    const membership = await getMembership(userId, organizationId, db)
+    if (!membership) return false
+    const [user] = await db
+      .select({ userType: schema.User.userType })
+      .from(schema.User)
+      .where(eq(schema.User.id, userId))
+      .limit(1)
+    return user?.userType === 'USER'
+  }
+
+  // Use org cache — zero DB queries
+  const { members } = await getOrgCache().getOrRecompute(organizationId, ['members'])
+  const member = members.find((m) => m.userId === userId)
+  return !!member && member.user?.userType === 'USER'
+}
+
+/** Count of ACTIVE members in an organization. */
+export async function getActiveMemberCount(
+  organizationId: string,
+  db: Database = defaultDb
+): Promise<number> {
+  const [row] = await db
+    .select({ value: count() })
+    .from(schema.OrganizationMember)
+    .where(
+      and(
+        eq(schema.OrganizationMember.organizationId, organizationId),
+        eq(schema.OrganizationMember.status, 'ACTIVE')
+      )
+    )
+  return row?.value ?? 0
+}
+
+/**
+ * All members of an organization (excludes system users), sorted by role then
+ * name then email.
+ */
+export async function getOrganizationMembers(organizationId: string, db: Database = defaultDb) {
+  const members = await db.query.OrganizationMember.findMany({
+    where: eq(schema.OrganizationMember.organizationId, organizationId),
+    with: {
+      user: {
+        columns: {
+          id: true,
+          name: true,
+          email: true,
+          image: true,
+        },
+        // Exclude system users from member lists
+        where: (user: typeof schema.User, { eq }) => eq(user.userType, 'USER'),
+      },
+    },
+    orderBy: [asc(schema.OrganizationMember.role)],
+  })
+
+  const roleOrder: Record<OrganizationRole, number> = { OWNER: 0, ADMIN: 1, USER: 2 }
+  return members.sort((a, b) => {
+    const roleDifference = roleOrder[a.role] - roleOrder[b.role]
+    if (roleDifference !== 0) {
+      return roleDifference
+    }
+
+    const nameA = a.user?.name?.toLocaleLowerCase() ?? ''
+    const nameB = b.user?.name?.toLocaleLowerCase() ?? ''
+    if (nameA !== nameB) {
+      return nameA.localeCompare(nameB)
+    }
+
+    const emailA = a.user?.email?.toLocaleLowerCase() ?? ''
+    const emailB = b.user?.email?.toLocaleLowerCase() ?? ''
+    return emailA.localeCompare(emailB)
+  })
 }
 
 /** Direct DB query — used inside transactions where cache reads are unsafe */
